Rename Gaussian and Mean files
