+++ piff/gaussian_model.py
